+++ lib/src/validation.dart
@@ -1,12 +1,20 @@
-library web_scraper;
+/*
+  Developed by Tushar Ojha
+  GitHub: https://github.com/tusharojha
+  Twitter: https://twitter.com/techtusharojha
+  Feel free to improve the web_scraper library.
+*/
 
+/// Validation Class containing all functions related to URL validation.
 class Validation {
   static final _ipv6 =
       RegExp(r'^::|^::1|^([a-fA-F0-9]{1,4}::?){1,7}([a-fA-F0-9]{1,4})$');
   static final _ipv4Maybe =
       RegExp(r'^(\d?\d?\d)\.(\d?\d?\d)\.(\d?\d?\d)\.(\d?\d?\d)$');
 
+  /// The isBaseURL function checks the base URLs like https://pub.dev.
   ValidationReturn isBaseURL(String str) {
+    // Protocols supported for web scraping includes http & https.
     const protocols = ['http', 'https'];
     var protocol, split, host;
 
@@ -111,6 +119,8 @@ class Validation {
   }
 }
 
+/// ValidationReturn class provides the result of validation
+/// including [isCorrect] and [description] for more details.
 class ValidationReturn {
   bool isCorrect;
   String description;