+++ lib/web_scraper.dart
@@ -1,23 +1,31 @@
 /*
   Developed by Tushar Ojha
   GitHub: https://github.com/tusharojha
-  Twitter: https://twitter.com/tusharojha
+  Twitter: https://twitter.com/techtusharojha
   Feel free to improve the web_scraper library.
 */
 
+/// Web Scraper Library
+///
+/// A very basic web scraper implementation to scrap html elements from a web page.
+///
+/// Package supports for both dart & flutter.
+///
+/// Contributions are welcome at https://github.com/tusharojha/web_scraper
 library web_scraper;
 
 import 'dart:async';
 
+import 'package:html/dom.dart';
 import 'package:html/parser.dart'; // Contains HTML parsers to generate a Document object.
 import 'package:http/http.dart'; // Contains a client for making API calls.
 
-import 'validation.dart';
+import 'src/validation.dart'; // Contains validation functions for URLs
 
 /// WebScraper Main Class.
 class WebScraper {
-  // Response Object of web scrapping the website.
-  var _response;
+  // Parsed document from the response inside the try/catch of the loadWebPage() method.
+  Document _document;
 
   // Time elapsed in loading in milliseconds.
   int timeElaspsed;
@@ -25,30 +33,32 @@ class WebScraper {
   // Base url of the website to be scrapped.
   String baseUrl;
 
-  Map<int, Client> Clients;
-
   /// Creates the web scraper instance.
-  WebScraper(String baseUrl) {
-    var v = Validation().isBaseURL(baseUrl);
-    if (!v.isCorrect) {
-      throw WebScraperException(v.description);
+  WebScraper([String baseUrl]) {
+    if (baseUrl != null) {
+      var v = Validation().isBaseURL(baseUrl);
+      if (!v.isCorrect) {
+        throw WebScraperException(v.description);
+      }
+      this.baseUrl = baseUrl;
     }
-    this.baseUrl = baseUrl;
   }
 
   /// Loads the webpage into response object.
   Future<bool> loadWebPage(String route) async {
-    if (baseUrl != null || baseUrl != '') {
+    if (baseUrl != null && baseUrl != '') {
       final stopwatch = Stopwatch()..start();
       var client = Client();
 
       try {
-        _response = await client.get(baseUrl + route);
+        var _response = await client.get(baseUrl + route);
         // Calculating Time Elapsed using timer from dart:core.
         if (_response != null) {
           timeElaspsed = stopwatch.elapsed.inMilliseconds;
           stopwatch.stop();
           stopwatch.reset();
+          // Parses the response body once it's retrieved to be used on the other methods.
+          _document = parse(_response.body);
         }
       } catch (e) {
         throw WebScraperException(e.message);
@@ -58,14 +68,43 @@ class WebScraper {
     return false;
   }
 
+  /// Loads the webpage URL into response object without requiring the two-step process of base + route.
+  /// Unlike the the two-step process, the URL is NOT validated before being requested.
+  Future<bool> loadFullURL(String page) async {
+    var client = Client();
+    try {
+      var _response = await client.get(page);
+      // Calculating Time Elapsed using timer from dart:core.
+      if (_response != null) {
+        // Parses the response body once it's retrieved to be used on the other methods.
+        _document = parse(_response.body);
+      }
+    } catch (e) {
+      throw WebScraperException(e.message);
+    }
+    return true;
+  }
+
+  /// Loads a webpage that was previously loaded and stored as a String by using [getPageContent].
+  /// This operation is synchronous and returns a true bool once the string has been loaded and is ready to
+  /// be queried by either [getElement], [getElementTitle] or [getElementAttribute].
+  bool loadFromString(String responseBodyAsString) {
+    try {
+      // Parses the response body once it's retrieved to be used on the other methods.
+      _document = parse(responseBodyAsString);
+    } catch (e) {
+      throw WebScraperException(e.message);
+    }
+    return true;
+  }
+
   /// Returns the list of all data enclosed in script tags of the document.
   List<String> getAllScripts() {
-    // The _response should not be null (loadWebPage must be called before getAllScripts).
-    assert(_response != null);
-    var document = parse(_response.body);
+    // The _document should not be null (loadWebPage must be called before getAllScripts).
+    assert(_document != null);
 
     // Quering the list of elements by tag names.
-    var scripts = document.getElementsByTagName('script');
+    var scripts = _document.getElementsByTagName('script');
     var result = <String>[];
 
     // Looping in all script tags of the document.
@@ -79,17 +118,17 @@ class WebScraper {
   }
 
   /// Returns Map between given variable names and list of their occurence in the script tags
-  /// ex. if document contains
-  /// <script> var a = 15; var b = 10; </script>
-  /// <script> var a = 9; </script>
-  /// method will return {a: ['var a = 15;', 'var a = 9;'], b: ['var b = 10;'] }.
+  ///
+  // ex. if document contains
+  // <script> var a = 15; var b = 10; </script>
+  // <script> var a = 9; </script>
+  // method will return {a: ['var a = 15;', 'var a = 9;'], b: ['var b = 10;'] }.
   Map<String, dynamic> getScriptVariables(List<String> variableNames) {
-    // The _response should not be null (loadWebPage must be called before getScriptVariables).
-    assert(_response != null);
-    var document = parse(_response.body);
+    // The _document should not be null (loadWebPage must be called before getScriptVariables).
+    assert(_document != null);
 
     // Quering the list of elements by tag names.
-    var scripts = document.getElementsByTagName('script');
+    var scripts = _document.getElementsByTagName('script');
 
     var result = <String, List<String>>{};
 
@@ -97,7 +136,7 @@ class WebScraper {
     for (var script in scripts) {
       // Looping in all the variable names that are required to extract.
       for (var variableName in variableNames) {
-        // regular expression to get the variable names
+        // Regular expression to get the variable names.
         var re = RegExp(
             '$variableName *=.*?;(?=([^\"\']*\"[^\"\']*\")*[^\"\']*\$)',
             multiLine: true);
@@ -105,7 +144,7 @@ class WebScraper {
         Iterable matches = re.allMatches(script.text);
         matches.forEach((match) {
           if (match != null) {
-            // list for all the occurence of the variable name
+            // List for all the occurence of the variable name.
             var temp = result[variableName];
             if (result[variableName] == null) {
               temp = [];
@@ -122,29 +161,91 @@ class WebScraper {
   }
 
   /// Returns webpage's html in string format.
-  String getPageContent() => _response != null
-      ? _response.body.toString()
+  String getPageContent() => _document != null
+      ? _document.outerHtml
       : throw WebScraperException(
           'ERROR: Webpage need to be loaded first, try calling loadWebPage');
 
-  /// Returns List of elements found at specified address.
+  /// Returns List of elements titles found at specified address.
   /// Example address: "div.item > a.title" where item and title are class names of div and a tag respectively.
-  List<Map<String, dynamic>> getElement(String address, List<String> attribs) {
-    // Attribs are the list of attributes required to extract from the html tag(s) ex. ['href', 'title'].
-    if (_response == null) {
+  /// For ease of access, when using Chrome inspection tool, right click the item you want to copy, then click "Inspect" and at the console, right click the highlighted item, right click and then click "Copy > Copy selector" and provide as String address parameter to this method.
+  List<String> getElementTitle(String address) {
+    if (_document == null) {
       throw WebScraperException(
           'getElement cannot be called before loadWebPage');
     }
-    // Using html parser and query selector to get a list of particular element.
-    var document = parse(_response.body);
-    var elements = document.querySelectorAll(address);
+    // Using query selector to get a list of particular element.
+    var elements = _document.querySelectorAll(address);
+
+    var elementData = <String>[];
+
+    for (var element in elements) {
+      // Checks if the element's text is null before adding it to the list.
+      if (element.text.trim() != '') {
+        elementData.add(element.text);
+      }
+    }
+    return elementData;
+  }
+
+  /// Returns List of elements' attributes found at specified address respecting the provided attribute requirement.
+  ///
+  /// Example address: "div.item > a.title" where item and title are class names of div and a tag respectively.
+  /// For ease of access, when using Chrome inspection tool, right click the item you want to copy, then click "Inspect" and at the console, right click the highlighted item, right click and then click "Copy > Copy selector" and provide as String parameter to this method.
+  /// Attributes are the bits of information between the HTML tags.
+  /// For example in <div class="strong and bold" style="width: 100%;" title="Fierce!">
+  /// The element would be "div.strong.and.bold" and the possible attributes to fetch would be EIHER "style" OR "title" returning with EITHER of the values "width: 100%;" OR "Fierce!" respectively.
+  /// To retrieve multiple attributes at once from a single element, please use getElement() instead.
+  List<String> getElementAttribute(String address, String attrib) {
+    // Attribs are the list of attributes required to extract from the html tag(s) ex. ['href', 'title'].
+    if (_document == null) {
+      throw WebScraperException(
+          'getElement cannot be called before loadWebPage');
+    }
+    // Using query selector to get a list of particular element.
+    var elements = _document.querySelectorAll(address);
+    // ignore: omit_local_variable_types
+    List<String> elementData = [];
+
+    for (var element in elements) {
+      var attribData = <String, dynamic>{};
+      attribData[attrib] = element.attributes[attrib];
+      // Checks if the element's attribute is null before adding it to the list.
+      if (attribData[attrib] != null) {
+        elementData.add(attribData[attrib]);
+      }
+    }
+    return elementData;
+  }
+
+  /// Returns List of elements found at specified address.
+  /// Example address: "div.item > a.title" where item and title are class names of div and a tag respectively.
+  ///
+  /// Sometimes the last address is not present consistently throughout the webpage. Use "extraAddress" to catch its attributes.
+  /// Example extraAddress: "a"
+  List<Map<String, dynamic>> getElement(String address, List<String> attribs,
+      {String extraAddress}) {
+    // Attribs are the list of attributes required to extract from the html tag(s) ex. ['href', 'title'].
+    if (_document == null) {
+      throw WebScraperException(
+          'getElement cannot be called before loadWebPage');
+    }
+    // Using query selector to get a list of particular element.
+    var elements = _document.querySelectorAll(address);
     // ignore: omit_local_variable_types
     List<Map<String, dynamic>> elementData = [];
 
     for (var element in elements) {
       var attribData = <String, dynamic>{};
       for (var attrib in attribs) {
-        attribData[attrib] = element.attributes[attrib];
+        if (extraAddress != null) {
+          var extraElement = element.querySelector(extraAddress);
+          if (extraElement != null) {
+            attribData[attrib] = extraElement.attributes[attrib];
+          }
+        } else {
+          attribData[attrib] = element.attributes[attrib];
+        }
       }
       elementData.add({
         'title': element.text,